User Authorization

+++ src/Ilios/CoreBundle/Controller/CurriculumInventorySequenceController.php
@@ -33,9 +33,9 @@ class CurriculumInventorySequenceController extends FOSRestController
      *   resource = true,
      *   requirements={
      *     {
-     *        "name"="report",
-     *        "dataType"="",
-     *        "requirement"="",
+     *        "name"="id",
+     *        "dataType"="integer",
+     *        "requirement"="\d+",
      *        "description"="CurriculumInventorySequence identifier."
      *     }
      *   },
@@ -54,7 +54,14 @@ class CurriculumInventorySequenceController extends FOSRestController
      */
     public function getAction($id)
     {
-        $answer['curriculumInventorySequences'][] = $this->getOr404($id);
+        $curriculumInventorySequence = $this->getOr404($id);
+
+        $authChecker = $this->get('security.authorization_checker');
+        if (! $authChecker->isGranted('view', $curriculumInventorySequence)) {
+            throw $this->createAccessDeniedException('Unauthorized access!');
+        }
+
+        $answer['curriculumInventorySequences'][] = $curriculumInventorySequence;
 
         return $answer;
     }
@@ -126,6 +133,11 @@ class CurriculumInventorySequenceController extends FOSRestController
                 $offset
             );
 
+        $authChecker = $this->get('security.authorization_checker');
+        $result = array_filter($result, function ($entity) use ($authChecker) {
+            return $authChecker->isGranted('view', $entity);
+        });
+
         //If there are no matches return an empty array
         $answer['curriculumInventorySequences'] =
             $result ? $result : new ArrayCollection([]);
@@ -158,9 +170,22 @@ class CurriculumInventorySequenceController extends FOSRestController
     public function postAction(Request $request)
     {
         try {
-            $new  =  $this->getCurriculumInventorySequenceHandler()
-                ->post($this->getPostData($request));
-            $answer['curriculumInventorySequences'] = [$new];
+            $handler = $this->getCurriculumInventorySequenceHandler();
+
+            $curriculumInventorySequence = $handler->post($this->getPostData($request));
+
+            $authChecker = $this->get('security.authorization_checker');
+            if (! $authChecker->isGranted('create', $curriculumInventorySequence)) {
+                throw $this->createAccessDeniedException('Unauthorized access!');
+            }
+
+            $this->getCurriculumInventorySequenceHandler()->updateCurriculumInventorySequence(
+                $curriculumInventorySequence,
+                true,
+                false
+            );
+
+            $answer['curriculumInventorySequences'] = [$curriculumInventorySequence];
 
             $view = $this->view($answer, Codes::HTTP_CREATED);
 
@@ -198,7 +223,7 @@ class CurriculumInventorySequenceController extends FOSRestController
     {
         try {
             $curriculumInventorySequence = $this->getCurriculumInventorySequenceHandler()
-                ->findCurriculumInventorySequenceBy(['report'=> $id]);
+                ->findCurriculumInventorySequenceBy(['id'=> $id]);
             if ($curriculumInventorySequence) {
                 $code = Codes::HTTP_OK;
             } else {
@@ -207,11 +232,26 @@ class CurriculumInventorySequenceController extends FOSRestController
                 $code = Codes::HTTP_CREATED;
             }
 
-            $answer['curriculumInventorySequence'] =
-                $this->getCurriculumInventorySequenceHandler()->put(
-                    $curriculumInventorySequence,
-                    $this->getPostData($request)
-                );
+            $handler = $this->getCurriculumInventorySequenceHandler();
+
+            $curriculumInventorySequence = $handler->put(
+                $curriculumInventorySequence,
+                $this->getPostData($request)
+            );
+
+            $authChecker = $this->get('security.authorization_checker');
+            if (! $authChecker->isGranted('edit', $curriculumInventorySequence)) {
+                throw $this->createAccessDeniedException('Unauthorized access!');
+            }
+
+            $this->getCurriculumInventorySequenceHandler()->updateCurriculumInventorySequence(
+                $curriculumInventorySequence,
+                true,
+                true
+            );
+
+            $answer['curriculumInventorySequence'] = $curriculumInventorySequence;
+
         } catch (InvalidFormException $exception) {
             return $exception->getForm();
         }
@@ -230,9 +270,9 @@ class CurriculumInventorySequenceController extends FOSRestController
      *   resource = true,
      *   requirements={
      *     {
-     *         "name" = "report",
-     *         "dataType" = "",
-     *         "requirement" = "",
+     *         "name" = "id",
+     *         "dataType" = "integer",
+     *         "requirement" = "\d+",
      *         "description" = "CurriculumInventorySequence identifier"
      *     }
      *   },
@@ -254,6 +294,11 @@ class CurriculumInventorySequenceController extends FOSRestController
     {
         $curriculumInventorySequence = $this->getOr404($id);
 
+        $authChecker = $this->get('security.authorization_checker');
+        if (! $authChecker->isGranted('delete', $curriculumInventorySequence)) {
+            throw $this->createAccessDeniedException('Unauthorized access!');
+        }
+
         try {
             $this->getCurriculumInventorySequenceHandler()
                 ->deleteCurriculumInventorySequence($curriculumInventorySequence);
@@ -273,7 +318,7 @@ class CurriculumInventorySequenceController extends FOSRestController
     protected function getOr404($id)
     {
         $curriculumInventorySequence = $this->getCurriculumInventorySequenceHandler()
-            ->findCurriculumInventorySequenceBy(['report' => $id]);
+            ->findCurriculumInventorySequenceBy(['id' => $id]);
         if (!$curriculumInventorySequence) {
             throw new NotFoundHttpException(sprintf('The resource \'%s\' was not found.', $id));
         }
